+++ app/imports/ui/components/standard-subcards/modal/actions-subcards/actions-subcards.html
@@ -1,0 +1,6 @@
+<template name="Actions_Subcards">
+  {{#with action}}
+    {{> Subcards_Notes_Edit
+        notes=this.notes}}
+  {{/with}}
+</template>
