+++ app/imports/ui/components/standard-subcards/modal/actions-subcards/index.js
@@ -1,0 +1,2 @@
+import './actions-subcards.html';
+import './actions-subcards.js';
