+++ app/imports/ui/components/standard-subcards/modal/index.js
@@ -1,2 +1,7 @@
 import './modal.html';
 import './modal.js';
+
+import './nc-subcards';
+import './standards-subcards';
+import './risks-subcards';
+import './actions-subcards';
