+++ app/imports/ui/components/standard-subcards/modal/nc-subcards/index.js
@@ -1,0 +1,2 @@
+import './nc-subcards.html';
+import './nc-subcards.js';
