+++ app/imports/ui/components/standard-subcards/modal/nc-subcards/nc-subcards.html
@@ -1,0 +1,39 @@
+<template name="NC_Subcards">
+  {{! NON-CONFORMITIES SUBCARDS}}
+  {{#with NC}}
+    {{> Subcards_Notify_Edit
+        doc=this
+        documentType="non-conformity"
+        onUpdate=onUpdateNotifyUserCb}}
+
+    {{> Subcards_Occurrences_Edit
+        _id=this._id
+        occurrencesIds=this.occurrences}} {{! can`t find occurences. Where are they?}}
+
+    {{> Subcards_Actions_Edit
+        type="CA"
+        documentId=this._id
+        documentType="non-conformity"}}
+
+    {{> Subcards_Actions_Edit
+        type="PA"
+        documentId=this._id
+        documentType="non-conformity"}}
+
+    {{> Subcards_ImprovementPlan_Edit
+        documentId=this._id
+        documentType="non-conformity"}}
+
+    {{> Subcards_LessonsLearned_Edit
+        documentId=this._id
+        documentType="non-conformity"
+        linkedTo=this.title
+        linkedToId=this.sequentialId}}
+
+    {{> Subcards_OtherFiles_Edit
+        _id=this._id
+        files=this.files
+        uploaderMetaContext=uploaderMetaContext
+        slingshotDirective=slingshotDirective}}
+  {{/with}}
+</template>
