+++ app/imports/ui/components/standard-subcards/modal/risks-subcards/index.js
@@ -1,0 +1,2 @@
+import './risks-subcards.html';
+import './risks-subcards.js';
