+++ app/imports/ui/components/standard-subcards/modal/risks-subcards/risks-subcards.html
@@ -1,0 +1,14 @@
+<template name="Risks_Subcards">
+  {{! RISKS SUBCARDS}}
+  {{#with risk}}
+    {{> Subcards_RiskScoring_Edit
+        _id=this._id
+        score=this.score}}
+
+    {{> Subcards_ImprovementPlan_Edit
+        _id=this._id
+        documentType="risk"
+        label="Treatment plan"
+        treatmentPlan=this.treatmentPlan}}
+  {{/with}}
+</template>
