+++ app/imports/ui/components/standard-subcards/modal/standards-subcards/index.js
@@ -1,0 +1,2 @@
+import './standards-subcards.html';
+import './standards-subcards.js';
