+++ app/imports/ui/components/standard-subcards/modal/standards-subcards/standards-subcards.html
@@ -1,0 +1,12 @@
+<template name="Standards_Subcards">
+  {{! STANDARDS SUBCARDS}}
+  {{#with standard}}
+    {{> Subcards_NonConformities_Edit
+      _id=this._id
+      _query=_getNCsQuery}}
+
+  {{> Subcards_Actions_Edit
+      type="RC"
+      standardId=this._id}}
+  {{/with}}
+</template>
